Reimplemented user authorization module.
General concept remained the same but the token has changed. Now it's cryptographically signed message containing ID of a user object (UID) which indetifies a user in the database. This is different from previously used server-side mapping 'token -> UID'. New token is valid for a given period of time and so should be refreshed if needed. This should be done in the background, not visible for user.