Implemented POST of new recipe.
Each tag should be given seperately, eg. ?tag=1&tag=blabla
Tags can be given either as a string or an int. In the first case they are matched by name, in second an attempt is made to match tag with an id but if it fails match by name is attempted. All in all, if a tag is not found, new one will be created.
Ingredients should be sent as a base64-encoded JSON's list of lists (not tuples in JSON): [[ingredient name, amount, unit name], ...]
Photos attached to new recipe must not be already attached to any other recipe.